requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asyncpg==0.30.0
attrs==25.4.0
auth0-ai==0.2.0
auth0-ai-langchain==0.2.0
auth0-python==4.13.0
Authlib==1.2.1
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
cryptography==46.0.3
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-environ==0.10.0
ecdsa==0.19.1
et_xmlfile==2.0.0
filetype==1.2.0
frozenlist==1.8.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-postgres==0.0.15
langchain-text-splitters==0.3.11
langgraph==0.3.34
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.74
langsmith==0.4.38
lxml==6.0.2
Markdown==3.9
marshmallow==3.26.1
msg-parser==1.2.0
multidict==6.7.0
mypy_extensions==1.1.0
nanoid==2.0.0
numpy==2.3.4
olefile==0.47
openai==2.6.0
openfga-sdk==0.9.7
openpyxl==3.1.5
opentelemetry-api==1.38.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20221105
pgvector==0.3.6
pillow==12.0.0
postgrest==2.22.2
propcache==0.4.1
psycopg==3.2.11
psycopg-pool==3.2.6
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
pypandoc==1.15
pypdf==3.16.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.3.0
python-magic==0.4.27
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
realtime==2.22.2
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
storage3==2.22.2
StrEnum==0.4.15
supabase-auth==2.22.2
supabase-functions==2.22.2
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
whitenoise==6.11.0
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
